restructured files and eat button functionality
