+++ src/javascripts/components/eat/eat.scss
@@ -1,0 +1,8 @@
+@import '../../../styles/variables';
+
+#eat {
+  width: $quad-width;
+  background-color: pink;
+  height: $quad-height;
+}
+
