+++ src/javascripts/components/fight/fight.scss
@@ -1,0 +1,7 @@
+@import '../../../styles/variables';
+
+#fight {
+  width: $quad-width;
+  height: $quad-height;
+  background-color: rgb(243, 51, 51);
+}
