+++ src/javascripts/components/play/play.js
@@ -1,4 +1,6 @@
-import utilities from '../helpers/utilities';
+import utilities from '../../helpers/utilities';
+
+import './play.scss';
 
 const createPlayButtons = () => {
   let domString = '';
@@ -7,9 +9,11 @@ const createPlayButtons = () => {
     <button id='super-fun' class='button'>Super fun!</button>
     <button id='fun' class='button'>Fun</button>
   </div>
-  <div id="progress-bar">
-      <progress value='1' label="fullness"></progress>
-  </div>
+  <div id="progress-container">
+      <div id="myProgress">
+        <div id="myBar" style="width: 20%"></div>
+      </div>
+    </div>
   `;
   utilities.printToDom(domString, 'play');
 };
