+++ src/javascripts/components/play/play.scss
@@ -1,0 +1,7 @@
+@import '../../../styles/variables';
+
+#play {
+    width: $quad-width;
+    height: $quad-height;
+    background-color: rgb(7, 212, 110);
+}
