+++ src/javascripts/components/sleep/sleep.js
@@ -1,4 +1,7 @@
-import utilities from '../helpers/utilities';
+import utilities from '../../helpers/utilities';
+
+import './sleep.scss';
+
 
 const createSleepButtons = () => {
   let domString = '';
@@ -7,9 +10,11 @@ const createSleepButtons = () => {
     <button id='nap' class='button'>Nap</button>
     <button id='deep-sleep' class='button'>Deep-sleep</button>
   </div>
-  <div id="progress-bar">
-      <progress value='1' label="fullness"></progress>
-  </div>
+  <div id="progress-container">
+      <div id="myProgress">
+        <div id="myBar" style="width: 80%"></div>
+      </div>
+    </div>
   `;
   utilities.printToDom(domString, 'sleep');
 };
