+++ src/javascripts/components/sleep/sleep.scss
@@ -1,0 +1,8 @@
+@import '../../../styles/variables';
+
+
+#sleep {
+    width: $quad-width;
+    height: $quad-height;
+    background-color: lightblue;
+}
