+++ src/javascripts/main.js
@@ -1,9 +1,9 @@
 import '../styles/main.scss';
 import 'bootstrap';
-import eat from './components/eat';
-import play from './components/play';
-import fight from './components/fight';
-import sleep from './components/sleep';
+import eat from './components/eat/eat';
+import play from './components/play/play';
+import fight from './components/fight/fight';
+import sleep from './components/sleep/sleep';
 
 
 const init = () => {
