+++ src/styles/_variables.scss
@@ -1,0 +1,7 @@
+$quad-width: 50%;
+$quad-height: 400px;
+$border-radius: 6px;
+$button-width: 150px;
+$button-height: 40px;
+$bar-width: 200px;
+$box-shadow: 5px 5px 10px 0px white;
