+++ src/styles/main.scss
@@ -1,10 +1,6 @@
 @import "~bootstrap/scss/bootstrap";
+@import './variables';
 
-$quad-width: 50%;
-$quad-height: 400px;
-$border-radius: 6px;
-$button-width: 150px;
-$button-height: 40px;
 
 body {
     background-color: blue;
@@ -16,30 +12,6 @@ body {
     justify-content: center;
   }
 
-  #eat {
-    width: $quad-width;
-    background-color: pink;
-    height: $quad-height;
-  }
-
-  #play {
-    width: $quad-width;
-    height: $quad-height;
-    background-color: rgb(7, 212, 110);
-  }
-
-  #fight {
-    width: $quad-width;
-    height: $quad-height;
-    background-color: rgb(243, 51, 51);
-  }
-
-  #sleep {
-    width: $quad-width;
-    height: $quad-height;
-    background-color: lightblue;
-  }
-
   .button {
     width: $button-width;
     height: $button-height;
@@ -48,7 +20,7 @@ body {
     background-color: rgb(22, 67, 95);
     color: white;
     border: none;
-    box-shadow: 5px 5px 10px 0px #FFF;
+    box-shadow: $box-shadow;
   }
 
   .button-container {
@@ -56,11 +28,29 @@ body {
     justify-content: space-between;
   }
 
-  #progress-bar {
+   #progress-container {
     display: flex;
     justify-content: center;
+  
+  
+    #myProgress {
+      width: $bar-width;
+      background-color: rgb(190, 190, 201);
+      border-radius: $border-radius;
+      box-shadow: 5px 5px 10px 0px;
+      
+  
+    
+      #myBar {
+        height: 15px;
+        background-color: rgb(74, 136, 207);
+        text-align: center;
+        line-height: 30px;
+        color: white;
+        border-radius: $border-radius;
+      }
+    }
   }
+  
 
-  progress {
-    height: 60px;
-  }
+
